single threaded

+++ import-graph.py
@@ -4,12 +4,13 @@ import sys
 import requests
 import json
 import bz2
-from multiprocessing import Pool, Manager
+from multiprocessing import Pool, Process, Manager
 from functools import partial
 import pickle
 from xml.etree import ElementTree
 import re
 import html
+from tqdm import tqdm
 
 DUMP_LANG = 'en'
 DUMP_DATE = '20200401'
@@ -24,7 +25,14 @@ RESULTS_DIR = 'results/'
 if not path.isdir(RESULTS_DIR): 
     os.mkdir(RESULTS_DIR)
 
-def collect_data_file(page_ids, graph, redirects, ifile, size):
+TEMP_DIR = 'temp/'
+if not path.isdir(TEMP_DIR): 
+    os.mkdir(TEMP_DIR)
+
+def collect_data_file(ifile, size):
+    if path.isfile(TEMP_DIR + ifile + '-graph.pkl') and path.isfile(TEMP_DIR + ifile + '-redirects.pkl'):
+        print('Skipped', ifile)
+        return
     if not path.isfile(DOWNLOAD_DIR + ifile) or not path.getsize(DOWNLOAD_DIR + ifile) == size:
         with open(DOWNLOAD_DIR + ifile, 'wb') as f:
             r = requests.get(DOWNLOAD_URL + ifile, stream=True)
@@ -48,6 +56,13 @@ def collect_data_file(page_ids, graph, redirects, ifile, size):
     with bz2.BZ2File(DOWNLOAD_DIR + ifile) as f:
         for l in f:
             l = l.decode('utf8').strip()
+
+            if l.startswith('<page'):
+                redirect_page = False
+                title = None
+                links = set()
+                continue
+
             if not redirect_page:
                 if l.startswith('<text'):
                     is_text = True
@@ -55,13 +70,19 @@ def collect_data_file(page_ids, graph, redirects, ifile, size):
                     continue
                 if l.endswith('</text>'):
                     is_text = False
+                    continue
                 if l.startswith('<title'):
                     title = html.unescape(l[7:-8]).lower()
+                    continue
                 if l.startswith('<redirect'):
                     redirect_page = True
                     to = html.unescape(l[17:-4]).lower()
                     if to in page_ids:
                         local_redirects[page_ids[title]] = page_ids[to]
+                    continue
+                if l.endswith('</page>'):
+                    local_graph[page_ids[title]] = list(links)
+                    continue
                 if is_text:
                     l = html.unescape(l.lower())
                     if not ignore_text:
@@ -70,46 +91,50 @@ def collect_data_file(page_ids, graph, redirects, ifile, size):
                         m = pattern.findall(l)
                         m = list(map(lambda s : s.split('|')[0].split('#')[0],m))
                         links.update(filter(lambda s : s in page_ids,m))
-                if l.endswith('</page>'):
-                    local_graph[page_ids[title]] = list(links)
-                
-            if l.startswith('<page'):
-                count += 1
-                redirect_page = False
-                title = None
-                links = set()
-    
 
-    graph.update(local_graph)
-    redirects.update(local_redirects)
+    
+    with open(TEMP_DIR + ifile + '-graph.pkl', 'wb') as f:
+        pickle.dump(local_graph, f)
+    
+    with open(TEMP_DIR + ifile + '-redirects.pkl', 'wb') as f:
+        pickle.dump(local_redirects, f)
+
     print('Finished', ifile, f"({count} pages)")
-            
 
-if __name__ == '__main__':
-    dumpstatus = requests.get(DOWNLOAD_URL + 'dumpstatus.json').json()
-    files = dumpstatus['jobs']['articlesmultistreamdump']['files']
-    data_files = sorted(list(filter(lambda f : 'index' not in f,files.keys())))
+if not path.isfile(DOWNLOAD_DIR + 'dumpstatus.json'):
+    with open(DOWNLOAD_DIR + 'dumpstatus.json', 'wb') as f:
+        r = requests.get(DOWNLOAD_URL + 'dumpstatus.json', stream=True)
+        f.writelines(r.iter_content(1024))
 
-    with open(RESULTS_DIR + 'index.pkl', 'rb') as f:
-        pi = pickle.load(f)
+with open(DOWNLOAD_DIR + 'dumpstatus.json') as f:
+    dumpstatus = json.load(f)
 
-    manager = Manager()
-    page_ids = manager.dict()
-    page_ids = pi
-    print('Index loaded')
+files = dumpstatus['jobs']['articlesmultistreamdump']['files']
+data_pages = sorted(list(filter(lambda f : 'index' not in f,files.keys())))
 
-    graph = manager.dict()
-    redirects = manager.dict()
+with open(RESULTS_DIR + 'index.pkl', 'rb') as f:
+    page_ids = pickle.load(f)
 
-    with Pool(4) as pool:
-        pool.starmap(partial(collect_data_file, page_ids, graph, redirects),[(i,files[i]['size']) for i in data_files])
+for i in data_pages:
+    collect_data_file(i,files[i]['size'])
 
-    with open(RESULTS_DIR + 'redirects.pkl', 'wb') as f:
-        d = redirects.copy()
-        print('Pickling', len(d), 'redirects')
-        pickle.dump(d, f)
-    
-    with open(RESULTS_DIR + 'raw_links.pkl', 'wb') as f:
-        d = graph.copy()
-        print('Pickling', len(d), 'pages of links')
-        pickle.dump(d, f)
+# merge
+graph = {}
+for i in os.listdir(TEMP_DIR):
+    if i.endswith('-graph.pkl'):
+        with open(i,'rb') as f:
+            d = pickle.load(f)
+            graph.update(d)
+with open(RESULTS_DIR + 'raw_graph.pkl','wb') as f:
+    pickle.dump(graph,f)
+del graph
+
+redirects = {}
+for i in os.listdir(TEMP_DIR):
+    if i.endswith('-redirects.pkl'):
+        with open(i,'rb') as f:
+            d = pickle.load(f)
+            redirects.update(d)
+with open(RESULTS_DIR + 'redirects.pkl','wb') as f:
+    pickle.dump(redirects,f)
+del redirects
